+++ PaxLab/Shot/src/main/java/dk/group6/shot/ShotSystem.java
@@ -20,16 +20,16 @@ import dk.group6.common.shot.ShotSPI;
 public class ShotSystem implements ShotSPI {
 
     @Override
-    public Entity shoot(int x, int y, float radian, World world) {
+    public Entity shoot(int x, int y, double radian, World world) {
         Shot shot = createShot(x, y, radian);
         world.addEntity(shot);
         return (Entity) shot;
     }
     
-    private Shot createShot(int x, int y, float radian){
+    private Shot createShot(int x, int y, double radian){
         Shot shot = new Shot();
         
-        shot.add(new PositionPart((float) x, (float) y));
+        shot.add(new PositionPart(x, y));
         shot.add(new MovingPart());
         SpritePart sp = new SpritePart("assets/spritshot.png", this.getClass());
         shot.add(sp);
